Fix loading previously generated batch images
The build_batch option when set to true destroys previously generated images, even if those images can be used again. I have added an if statement that checks to see if the batch directory for the dataset exists, if one doesn't then the computer creates the directory and generates new batch images. Otherwise, the computer will use the existing images in the directory.

However, there are two limitations to this method. If the computer is interrupted while it is building the batch images or the dataset is altered after the batch images have already been built, the batch image directory for that dataset must be deleted. If the directory is not deleted, it will result in a index error during training.

While this is not an ideal fix, it still saves quite a bit of time if training the same model with no modifications to dataset.